+++ .gitignore
@@ -1,0 +1,2 @@
+script/notify.py
+*.pickle
