+++ script/.ipynb_checkpoints/graph-checkpoint.dot
@@ -1,0 +1,1 @@
+digraph graphname{rankdir=TB;140698577020968 [label="BatchNormalization",fillcolor="#6495ED",style="filled",shape="record"];140698577055816 [label="gamma: (128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577019064 [label="MeanSquaredError",fillcolor="#6495ED",style="filled",shape="record"];140698577117424 [label="b: (128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577119552 [label="W: (128, 128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577021248 [label="BatchNormalization",fillcolor="#6495ED",style="filled",shape="record"];140698577021304 [label="(7, 6), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577117592 [label="(7, 4), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577021360 [label="(7, 128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577474000 [label="LinearFunction",fillcolor="#6495ED",style="filled",shape="record"];140698577056208 [label="gamma: (128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698576962056 [label="(7, 128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577021528 [label="(7, 128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577455760 [label="LinearFunction",fillcolor="#6495ED",style="filled",shape="record"];140698577117872 [label="(7, 128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577119944 [label="W: (128, 128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577433264 [label="LinearFunction",fillcolor="#6495ED",style="filled",shape="record"];140698577120056 [label="(7, 128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577118096 [label="(7, 4), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577433488 [label="gamma: (128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577118152 [label="W: (4, 128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577021976 [label="ReLU",fillcolor="#6495ED",style="filled",shape="record"];140698577120336 [label="b: (128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577120392 [label="W: (128, 6), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577120504 [label="(7, 128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577030392 [label="float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577436008 [label="BatchNormalization",fillcolor="#6495ED",style="filled",shape="record"];140698577118600 [label="b: (4), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577454472 [label="(7, 128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577057160 [label="beta: (128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577057328 [label="beta: (128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577020632 [label="LinearFunction",fillcolor="#6495ED",style="filled",shape="record"];140698577020688 [label="(7, 128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577057552 [label="beta: (128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577022760 [label="ReLU",fillcolor="#6495ED",style="filled",shape="record"];140698577020800 [label="ReLU",fillcolor="#6495ED",style="filled",shape="record"];140698577121176 [label="b: (128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577020856 [label="(7, 128), float32",fillcolor="#E0E0E0",style="filled",shape="octagon"];140698577020968 -> 140698577120056;140698577020856 -> 140698577020632;140698577120056 -> 140698577020800;140698577119552 -> 140698577474000;140698577020632 -> 140698577020688;140698577117424 -> 140698577433264;140698577120504 -> 140698577455760;140698577057160 -> 140698577436008;140698577433488 -> 140698577436008;140698577455760 -> 140698577118096;140698577021360 -> 140698577436008;140698577121176 -> 140698577474000;140698577057328 -> 140698577020968;140698577021248 -> 140698577454472;140698577117872 -> 140698577020968;140698577120392 -> 140698577433264;140698577021528 -> 140698577022760;140698577474000 -> 140698577117872;140698577454472 -> 140698577021976;140698577019064 -> 140698577030392;140698577118600 -> 140698577455760;140698577117592 -> 140698577019064;140698577021304 -> 140698577433264;140698577436008 -> 140698577021528;140698577433264 -> 140698577021360;140698577055816 -> 140698577020968;140698576962056 -> 140698577474000;140698577118096 -> 140698577019064;140698577119944 -> 140698577020632;140698577056208 -> 140698577021248;140698577118152 -> 140698577455760;140698577120336 -> 140698577020632;140698577020800 -> 140698577120504;140698577020688 -> 140698577021248;140698577057552 -> 140698577021248;140698577022760 -> 140698577020856;140698577021976 -> 140698576962056;}
