+++ script/.ipynb_checkpoints/graph2-checkpoint.dot
@@ -1,0 +1,1 @@
+digraph graphname{rankdir=TB;140375521125264 [fillcolor="#E0E0E0",shape="octagon",label="(32, 5), float32",style="filled"];}
