+++ script/dumy.py
@@ -1,0 +1,2 @@
+from notify import notify
+notify('test')
