+++ script/graph2.dot
@@ -1,0 +1,1 @@
+digraph graphname{rankdir=TB;140375516446848 [fillcolor="#E0E0E0",shape="octagon",label="(32, 5), float32",style="filled"];140375522279552 [fillcolor="#E0E0E0",shape="octagon",label="b: (128), float32",style="filled"];140375517362328 [fillcolor="#E0E0E0",shape="octagon",label="gamma: (128), float32",style="filled"];140375516297424 [fillcolor="#E0E0E0",shape="octagon",label="gamma: (128), float32",style="filled"];140375516449032 [fillcolor="#E0E0E0",shape="octagon",label="(32, 128), float32",style="filled"];140375516447016 [fillcolor="#E0E0E0",shape="octagon",label="(32, 128), float32",style="filled"];140375516449088 [fillcolor="#E0E0E0",shape="octagon",label="(32, 128), float32",style="filled"];140375517362552 [fillcolor="#E0E0E0",shape="octagon",label="gamma: (128), float32",style="filled"];140375516447296 [fillcolor="#E0E0E0",shape="octagon",label="(32, 128), float32",style="filled"];140375522280000 [fillcolor="#E0E0E0",shape="octagon",label="b: (128), float32",style="filled"];140375516500600 [fillcolor="#E0E0E0",shape="octagon",label="(32, 7), float32",style="filled"];140375516448752 [fillcolor="#6495ED",shape="record",label="ReLU",style="filled"];140375516500768 [fillcolor="#6495ED",shape="record",label="BatchNormalization",style="filled"];140375516449592 [fillcolor="#6495ED",shape="record",label="ReLU",style="filled"];140375516447576 [fillcolor="#E0E0E0",shape="octagon",label="(32, 128), float32",style="filled"];140375517363056 [fillcolor="#E0E0E0",shape="octagon",label="beta: (128), float32",style="filled"];140375516447632 [fillcolor="#E0E0E0",shape="octagon",label="(32, 128), float32",style="filled"];140375522282408 [fillcolor="#E0E0E0",shape="octagon",label="W: (128, 128), float32",style="filled"];140375516449704 [fillcolor="#6495ED",shape="record",label="LinearFunction",style="filled"];140375516447800 [fillcolor="#E0E0E0",shape="octagon",label="(32, 128), float32",style="filled"];140375517361320 [fillcolor="#E0E0E0",shape="octagon",label="W: (5, 128), float32",style="filled"];140375516448136 [fillcolor="#6495ED",shape="record",label="ReLU",style="filled"];140375521719744 [fillcolor="#E0E0E0",shape="octagon",label="beta: (128), float32",style="filled"];140375516448192 [fillcolor="#6495ED",shape="record",label="BatchNormalization",style="filled"];140375521850816 [fillcolor="#E0E0E0",shape="octagon",label="beta: (128), float32",style="filled"];140375522283024 [fillcolor="#E0E0E0",shape="octagon",label="b: (128), float32",style="filled"];140375516462664 [fillcolor="#6495ED",shape="record",label="LinearFunction",style="filled"];140375516448360 [fillcolor="#E0E0E0",shape="octagon",label="(32, 128), float32",style="filled"];140375522492144 [fillcolor="#E0E0E0",shape="octagon",label="W: (128, 7), float32",style="filled"];140375522283304 [fillcolor="#E0E0E0",shape="octagon",label="W: (128, 128), float32",style="filled"];140375516448584 [fillcolor="#E0E0E0",shape="octagon",label="(32, 128), float32",style="filled"];140375516448640 [fillcolor="#6495ED",shape="record",label="BatchNormalization",style="filled"];140375516479384 [fillcolor="#6495ED",shape="record",label="LinearFunction",style="filled"];140375516448696 [fillcolor="#6495ED",shape="record",label="LinearFunction",style="filled"];140375517362160 [fillcolor="#E0E0E0",shape="octagon",label="b: (5), float32",style="filled"];140375522282408 -> 140375516449704;140375522492144 -> 140375516479384;140375516448640 -> 140375516447296;140375516448136 -> 140375516447632;140375516500600 -> 140375516479384;140375516447296 -> 140375516448752;140375516448192 -> 140375516448360;140375516447800 -> 140375516448696;140375516449704 -> 140375516448584;140375516297424 -> 140375516448640;140375516447632 -> 140375516462664;140375517362552 -> 140375516448192;140375516479384 -> 140375516447016;140375517362160 -> 140375516448696;140375522280000 -> 140375516449704;140375516449088 -> 140375516448192;140375517362328 -> 140375516500768;140375516462664 -> 140375516449088;140375516449592 -> 140375516447576;140375521850816 -> 140375516448640;140375516447576 -> 140375516449704;140375522283304 -> 140375516462664;140375516500768 -> 140375516449032;140375517361320 -> 140375516448696;140375516447016 -> 140375516500768;140375516448696 -> 140375516446848;140375516449032 -> 140375516448136;140375522279552 -> 140375516462664;140375516448360 -> 140375516449592;140375522283024 -> 140375516479384;140375516448584 -> 140375516448640;140375521719744 -> 140375516448192;140375516448752 -> 140375516447800;140375517363056 -> 140375516500768;}
